araqne/issue#2018: fulltext search performance improve
- removed useless readInt call which can be slow(It doesn't use
  FileChannel)

+++ araqne-logstorage/src/main/java/org/araqne/logstorage/file/LogFileV3o.java
@@ -181,7 +181,7 @@ public class LogFileV3o extends FilePair<IndexBlockV3Header, LogFileV3o.RawDataB
 
 	@Override
 	public RawDataBlock getRawDataBlock(IndexBlockV3Header indexBlock) throws IOException {
-		ByteBuffer dataBuffer = ByteBuffer.allocate((int) indexBlock.getDataBlockLen());
+		ByteBuffer dataBuffer = ByteBuffer.allocate((int) indexBlock.getDataBlockLen().intValue());
 		StorageInputStream stream = getDataStream();
 		stream.seek(indexBlock.getPosOnData());
 		stream.readFully(dataBuffer.array());
